Create new marty instance for testing renderToString
